Separate XSLT compilation and rendering status
